requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
prometheus-client==0.19.0
tiktoken==0.5.2
python-dotenv==1.0.0
